fetch envvars via proc environ
